requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
pillow==10.2.0
python-dotenv==1.0.1